requirements.txt
beautifulsoup4==4.5.1
fbchat==0.4.0
lxml==3.6.4
PyYAML==3.12
requests==2.11.1
Yapsy==1.11.223
